broadcast: fix growth candidate query throwing on every tick
'p.subs1 > p.subs0 * $2' made Postgres infer the growth multiplier param ($2 =
1.5) as bigint from the bigint-multiply context, so it errored with 'invalid
input syntax for type bigint: 1.5' on EVERY tick (51x in logs). runGrowthStoryTick
threw before posting, so auto growth stories never fired — every growth_story on
the channel so far was a manual test send that bypasses this query. Cast the
multiplier to float8 (::float8), same class of param-type trap as the int4
overflow one. 336 candidates were sitting ready behind this.

+++ lib/broadcast/growth.ts
@@ -165,7 +165,12 @@ export async function runGrowthStoryTick(): Promise<GrowthTickResult> {
      )
      SELECT p.channel_id FROM pairs p
        JOIN niche_spy_channels c ON c.channel_id = p.channel_id
-      WHERE p.subs0 >= $1 AND p.subs1 > p.subs0 * $2
+      WHERE p.subs0 >= $1::bigint
+        -- Cast the multiplier to float8: 'bigint * $2' makes PG infer $2 as
+        -- bigint, and the fractional 1.5 fails to parse (the same param-type
+        -- trap as the int4 overflow one). This threw on EVERY tick, so auto
+        -- growth stories never fired — only manual test sends worked.
+        AND p.subs1::float8 > p.subs0::float8 * $2::float8
         AND ${NOT_AUTOGEN_SQL}
         AND NOT EXISTS (SELECT 1 FROM broadcast_posts b
                          WHERE b.featured_key = 'growth:'||p.channel_id
